moving components to components folder
